Use Tablayout for rudimentary settings
