+++ app/build.gradle
@@ -70,6 +70,7 @@ android {
 dependencies {
     testCompile 'junit:junit:4.12'
     compile 'com.android.support:appcompat-v7:22.0.0'
+    compile 'com.android.support:design:22.2.0'
     compile files('src/main/libs/neuroph-core-2.92.jar')
     compile files('src/main/libs/slf4j-android-1.7.21.jar')
     compile files('src/main/libs/slf4j-api-1.7.21.jar')
