+++ app/src/main/java/rtandroid/ballsort/gui/fragments/SettingsFragment.java
@@ -1,0 +1,45 @@
+package rtandroid.ballsort.gui.fragments;
+
+import android.support.v4.app.Fragment;
+import android.os.Bundle;
+import android.util.Log;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.Button;
+
+import rtandroid.ballsort.MainActivity;
+import rtandroid.ballsort.R;
+import rtandroid.ballsort.settings.SettingsManager;
+
+public class SettingsFragment extends Fragment
+{
+
+    @Override
+    public void onCreate(Bundle savedInstanceState)
+    {
+        super.onCreate(savedInstanceState);
+
+        Button btnSave = (Button) getView().findViewById(R.id.btnSave);
+        btnSave.setOnClickListener(v ->
+        {
+            SettingsManager.writeToFile();
+            Log.i(MainActivity.TAG, "Settings saved");
+        });
+
+        Button btnLoad = (Button) getView().findViewById(R.id.btnLoad);
+        btnLoad.setOnClickListener(v ->
+        {
+            SettingsManager.readFromFile();
+            Log.i(MainActivity.TAG, "Settings loaded");
+        });
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState)
+    {
+        View view = inflater.inflate(R.layout.fragment_settings, container, false);
+        return view;
+    }
+}
